automatically backed up by learn

+++ db/migrate/001_create_costumes.rb
@@ -12,7 +12,7 @@ class CreateCostumes < ActiveRecord::Migration
   def change
     create_table :costumes do |c|
       c.string :name
-      c.integer :price
+      c.float :price
       c.integer :size
       c.string :image_url
   end
